# source/maas-toolbox/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API service for managing tier-to-group mappings in the Open Data Hub Model as a Service (MaaS) project.\nThis API provides CRUD operations for managing tiers that map Kubernetes groups to user-defined subscription tiers.",
        "title": "Open Data Hub MaaS Toolbox API",
        "contact": {
            "name": "Bryon Baker",
            "url": "https://github.com/bryonbaker"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/groups/{group}/tiers": {
            "get": {
                "description": "Retrieve all tiers that contain the specified Kubernetes group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get tiers by group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tiers containing the group",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tier"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid group name format",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tiers": {
            "get": {
                "description": "Retrieve a list of all tiers in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "List all tiers",
                "responses": {
                    "200": {
                        "description": "List of tiers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tier"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tier with name, description, level, and groups. The tier name must be unique and cannot be changed after creation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Create a new tier",
                "parameters": [
                    {
                        "description": "Tier object",
                        "name": "tier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tier"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tier created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Tier"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - tier already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tiers/{name}": {
            "get": {
                "description": "Retrieve a tier by its name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Get a specific tier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tier name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tier details",
                        "schema": {
                            "$ref": "#/definitions/models.Tier"
                        }
                    },
                    "404": {
                        "description": "Tier not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a tier's description, level, or groups. The tier name cannot be changed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tiers"
                ],
                "summary": "Update a tier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tier name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tier update object (name field is ignored)",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated tier",
                        "schema": {
                            "$ref": "#/definitions/models.Tier"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tier not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a tier by its name",
                "tags": [
                    "tiers"
                ],
                "summary": "Delete a tier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tier name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content - tier deleted successfully"
                    },
                    "404": {
                        "description": "Tier not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tiers/{name}/groups": {
            "post": {
                "description": "Add a Kubernetes group to a tier. The group must not already exist in the tier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Add a group to a tier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tier name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Group to add",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated tier with new group",
                        "schema": {
                            "$ref": "#/definitions/models.Tier"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tier not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - group already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tiers/{name}/groups/{group}": {
            "delete": {
                "description": "Remove a Kubernetes group from a tier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Remove a group from a tier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tier name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group name to remove",
                        "name": "group",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated tier with group removed",
                        "schema": {
                            "$ref": "#/definitions/models.Tier"
                        }
                    },
                    "404": {
                        "description": "Tier or group not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AddGroupRequest": {
            "description": "Request body for adding a group to a tier",
            "type": "object",
            "required": [
                "group"
            ],
            "properties": {
                "group": {
                    "description": "Kubernetes group name to add",
                    "type": "string",
                    "example": "premium-users"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Tier": {
            "description": "Tier configuration that maps Kubernetes groups to a subscription tier",
            "type": "object",
            "properties": {
                "description": {
                    "description": "Tier description",
                    "type": "string",
                    "example": "Free tier for basic users"
                },
                "groups": {
                    "description": "List of Kubernetes groups",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "system:authenticated"
                    ]
                },
                "level": {
                    "description": "Tier level (non-negative integer)",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Tier name (immutable after creation)",
                    "type": "string",
                    "example": "free"
                }
            }
        }
    }
}